Deleted old version of object classifier and finalized the newest version
